<update> send info task

+++ game-server/src/main/java/nro/server/engine/entity/system/MapChangeSystem.java
@@ -102,7 +102,7 @@ public class MapChangeSystem extends IteratingSystem {
         // add entity vào area mới
         newArea.addEntity(client.getPlayerID());
 
-        pos.mapId = waypoint.getGoMap();
+        pos.mapId = (short) waypoint.getGoMap();
         pos.areaId = newArea.getInstanceId();
         pos.x = (short) xNew;
         pos.y = (short) yNew;
